Initial
I couldn't locate the repository where the aspell language packs
development takes place, hence this repository.

It stores extracted aspell-language tarballs in a way that should be
efficient for a git repository.

Instead of storing the tarballs we extract them into generic directories
that are overwritten everytime an updated language pack is added.